fix problem with retransmitting packets at the right time
The solution is to add a RetransmitEvent in the events queue with a delay equal to the given timeout duration in TCP. When the event is triggered it checks if the packet should be sent again or not. If it should be sent again then the it ads a RouteEvent to the events queue.