add and implement ChannelEvent
Channels are own events. This is practical because there should be some delay associated with sending packets over a channel.

Also it prevent duplicate code in the generating of new events. this is because all nodes are connected to channels in such a way that other than calling it self it should call the channel, and therfore not worry about what type of node the channel is connected to